flatpak-release.bst: Introduce the "extra" Flatpak branch
The branch name is in the format "<version>-extra".

It only includes the Mesa runtime extension at this time, but
additonal Flatpak images may be added to it in the future.

Extensions provided by both the regular and extra branches will take
precedence over one another based on their "priority" configuration
as defined in flatpak-metadata(5), allowing only one of them to be
mounted in the sandbox.

Distributors and users who aren't in a legal position to make use of
images included in the extra branch are required to filter them out
using a filter file as documented in flatpak-remote-add(1) and
flatpak-remote-modify(1).

Alternatively, the flatpak-mask(1) command can be used to prevent the
installation of Flatpak refs from this branch, like so:
$ flatpak mask *//*-extra

This addition requires the use of a separate flatpak_repo element
since the flatpak_repo plugin only supports the creation of
single-branch repositories at the moment.