flatpak-images: Add an additional version of the Mesa extension
This version of the Mesa extension includes extra video codecs, and
is provided as an option for users and distributors which are legally
positioned to make use of them.

This version of the extension is configured to take precendence over
the base Mesa extension if both are installed in the system. This is
achieved using the "priority" configuration as defined in
flatpak-metadata(5).