project.conf: Use a constant build-root
The default build-root in BuildStream contains the project name and
the path to the element in the project directory:
        build-root: /buildstream/%{project-name}/%{element-name}

Since the build directory can sometimes affect the build output itself
(as part of debug information, for example), having different build
directories for each element can result in cross-element
reproducibility issues for files that could otherwise be identical.

Making the build-root constant for all elements should prevent this.