include/strip.yml: Make the strip directory suffix configurable
The paths to the source and debug information directories are usually
embedded in debug information. This poses a problem when attempting
to achieve reproducible files across elements.

Unfortunately, paths to the source and debug information directories
of each element must be unique so that element artifcats could be
staged together without risk of file conflicts.

Nevertheless, specific elements that are never expected or meant to be
staged together, such as two alternative versions of the same element,
could use non-unique source and debug information directory paths
between them to gain cross-element reproducibility.

Making the directory suffix configurable allows achieving this on a
per-element basis.