bootstrap/include/common.yml: Always set the "bootstrapper" vendor
Before this change:
* Cross-compilation was done naturally when bootstrap_build_arch was
  different than target_arch because it resulted in the "arch" part
  of the triplet being different between --host and --build. This
  affected cross-compilers.
* Cross-compilation was not done naturally when bootstrap_build_arch
  was the same as target_arch, so the use of the "bootstrapper" vendor
  in the triplet was used to make --host and --build different and
  force cross-compilation. This affected the "normal" bootstrap.

Actually setting the different --host and --build parmameters to force
cross-compilation was done in bootstrap/include/target.yml, and is
only desired for elements that include it.

However, an oversight in that approach is that it also results in
every element which uses bootstrap/include/build.yml to get a
different cache key depending on whether it is built in the context
of the "normal" bootstrap or the cross-compilers bootstrap, and only
because each case uses a different vendor string ("unknown" vs.
"bootstrapper").

So instead, always use the "bootstrapper" vendor stringfor bootstrap.
It results in no functional change but it allows sharing cache keys
regardless of target_arch for elements that make use of
bootstrap/include/build.yml.