elements/bootstrap/build/base-sdk: Include bootstrap/include/build.yml
The value of the lib and libdir variables affects the cache key of
every element, import/stack/filter included. By default, the value of
lib, and therefore of libdir, depends on target_arch.

This mean that the cache key of every element in base-sdk, including
all elements that depend on them, is different when eleemnts build in
the context of a "normal" bootstrap or a cross-compilers bootstrap.

The addition of this include results in no functional change because
these elements don't make use of lib and libdir, but it allows sharing
cache keys for elements in bootstrap/build/base-sdk and
bootstrap/build regardless of target_arch.