Makefile: Align architecture handling to master
This is a partial backport of b7ee52213 to more easily share
architecture configuration between different VM targets.

The only change in behavior introduced is to allocate 2GB RAM
instead of 256MB for imageless VMs. This aligns imageless VMs RAM
configuration to the master branch and to the OSTree VM.