include/strip.yml: Add "debugdatadir" variable
The existing "debugdir" variable represents the runtime directory in
which debuginfo files are searched. Files in this directory are
expected to be located in the same relative path as their binary
counterparts under the /usr prefix.

The new "debugdatadir" variable represents the location in which
additional debug-related data resides, including source files and dwz
files. Files in this path don't have to conform to the directory
structure of the /usr prefix, both because they don't have binary
equivalents under /usr and also because their paths are embedded in
the debuginfo files themselves.

The "debugdatadir" variable can be used to embed source files and dwz
files in the artifact of Flatpak extensions.