flatpak-images/{platform,sdk}.bst: Add the org.freedesktop.Platform.GL.Debug extension point
It will be used to mount runtime extensions that contain debug
information for GL extensions.

It uses merge-dirs to create the same directory structure as the
org.freedesktop.Platform.GL extension. It does so under the
lib/debug%{libdir}/GL path, which gdb and other tools expect when
searching for debuginfo files for binaries.

It is enabled according to active-gl-driver so that it gets mounted
along with its org.freedesktop.Platform.GL extension counterpart
when installed. It is not installed automatically, however.

In addition, since the new extension point is located on a path shared
with the org.freedesktop.Sdk.Debug extension, the latter needs to have
a directory prepared so the two can be mounted at the same time.