ci: Explicitly tag update, cve and manifest jobs with aarch64
The last commit does not work as I intended, I seemingly forgot that a
runner needs to have *all* the tags listed in a job to pick it up. There
is an upstream GitLab issue to improve this (surprise, surprise), which
has seen some movement but likely won't land any time soon.

Given we're close to 21.08 EOL and these tasks tend to be fairly quick,
let's just throw them at the aarch64 runner for now.