Fix module signing and make linux.bst reproducible
Module signing was not reproducible since it was generated a temporary
key.

Also the external module signing was not doing anything because module
were compressed and not found.

Now by default, we do not sign the modules. Downstream can override
`components/linux-module-cert.bst` to provide the certificate. The
signing still happens outside of `linux.bst`.  But we do need to
insert the certificate in the kernel.

This will also allow building of external modules and load them
in lockdown mode (which is expected in secure boot).

We enable Machine Owner keys, so that machine owner can sign their own
modules and load them in secure boot.

The demo secure boot image is updated with the required lockdown
parameter.