linux: Make SND_HDA module only
Some HDA sound cards need quirks that are matched with specific
modules. If at initialization, if generic modules only are available,
the quirks will not be used. This issue happens in initramfs when the
modules are available and generic modules are built-in. To avoid this,
we need to make all snd-hda drivers modules.