systemd: Split out ukify and kernel-install
ukify is missing a python dependency to run properly. Instead of
adding that depdency to systemd, it is better to split it out, since
in the case of image-based OSes, it is more a build tool than runtime.

kernel-install is useful for our cases. So we can split it out in case
it is needed.