Add git-lfs
A lot of apps, generally IDEs and git clients depend on git-lfs to
provide functionality and some of them use the SDK as the runtime. So
add git-lfs to the SDK so that apps doesn't need to bundle this.

This isn't added to the runtime as runtime is missing git