Fix rust doc reproducibility issue
When rust builds its docs, it does so with many threads by default.  This causes
a race condition between `std`, `core`, `alloc` etc. crates on the creation of
the docs.  This gives inconsistent outputs in the documentation URLs for the
same file.

Instead, use a single thread in an attempt to remove the race.