Add shaderc in flatpak runtime
shaderc can be used at build-time by various apps. It also contains
shared libraries which can be linked to so they must be in Platform.

This makes shaderc libs checked for abi changes although we already
need to freeze everything on glslang abi changes so situation
wouldn't be worse in practice.

The alternative to this is apps continue to build pararel vulkan stack
because it's all interconected and adding one element at specific
version requires all other matching it.