linux: Disable PCIEASPM_POWER_SUPERSAVE
That does not work on my computer. Some PCIe devices are not
available. There is probably some power management to do. But when the
nvme boot disk does not appear during boot, there is not much to
do. It is better to switch back to BIOS configuration (which is
PCIEASPM_DEFAULT and enabled by default when PCIEASPM_POWER_* are not
enabled).