utils/generate_cve_report.py: Add tables of unversioned elements of git/archive sources
This extends the CVE report with additional tables for any unversioned
elements in the manifest, for 'git' and 'archive' source types,
indicating any CVE that relates to the individual element.

The tables can be used to identify buildstream elements which may
need their cpe version information set manually in the element, as
it cannot be determined from the source information.