vm/prepare-image.sh: Set root's initial password as unprovisioned
systemd-firstboot allows setting the root password from a credential.
Unfortunately, it only does that if root's shadow password is
!unprovisioned, which happens when creating the user with
systemd-sysusers. This is a valid representation of a locked password
even if the password isn't updated via systemd-firstboot.

A side effect of this change is that the image always has the
/etc/shadow and /etc/gshadow files with all relevant entries since they
get updated by systemd-sysusers. Previously only /etc/shadow was created
and was empty unless the root password was pre-configured with
--rootpassword.

Fixes: #1745