elements/components/ffmpeg.bst: ensure vvc decoder is always disabled
vvc decoder is implementation of H.266 standard with patent restrictions so we want to keep it disabled in free version. Even when that's the case right now before this change yet without explicit disabling it may be enabled in the future through dependency as it happened with h264 in the past.