ci: Fix order of if rules in check_mirror
if rules are evaluated in the order they appear until the first match
is found and after that the rest are ignored. We want the job to
always run on branches and not on tags.

So the tags rule should always be evaluated first as that will not
match on branches.

See https://docs.gitlab.com/ee/ci/yaml/#rules