mesa: Update gbm libdrm patch and rename
We do not need to force libdrm dep anymore as upstream did that in
24.2.1 but in 24.2.0 they also started linking libgbm directly with
libgallium_dri.

As this is built without any backends and the core mesa is in the
extension we can't link it with that.

Rename the patchfile to that effect. This patch must NOT be used for
mesa shipped in the extension.