components/fwupd.bst: Add missing libcbor dependency
This is a required runtime and  build dependency that was provided
only indirectly through systemd as a build-dependency. It's likely
that, under certain scenarios, these tools would end up broken due
to libcbor being missing as a runtime dependency.

This issue was caught by trying to build systemd without libfido2
which pulls libcbor in.