ca-certificates: Use a pinned copy of update-ca-trust script
This is done as the ca-cert element is a mis-mash of our packaging and
Fedora's scripts. A in the script by upstream that removes a symlink
creation and moves it to spec or removes it altogether requires follow-up
on our side.

Without a pinned copy, the build silently passes and we do not notice
anything.

This restores the compat symlinks in https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/issues/1781#note_2178827877
and `ca-bundle.trust.crt` https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/issues/1781#note_2176393763
as both were dropped by upstream without much explanation and at least
one of them has started to cause issues with Flathub apps.

See https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/issues/1781