bpi-f3: select downstream kernel and firmware based on new build option
Select the kernel downstream element instead of the upstream one based
on the new downstream_bsp option. This to avoid duplicating all the
elements depending on the kernel element. Also add the spacemit-k1
downstream firmware in vm/boot/efi/deps.bst as it needs to be included
in the initramfs to properly initialize hardware.