check-abi-mesa: Allow forward changes in extension libdrm
As platform libdrm is frozen it will have no forward change, but
extension libdrm is supposed to be freely updated. We only need to make
sure that extension libdrm does not break backwards compat with platform
libdrm but it is allowed to break forward compat.

Without --no-added-syms abdiff will exit with error even on forward
changes in extension libdrm. So pass --no-added-syms to suppress forward
changes