Add work-around to be able to setcap/getcap in Buildstream
This adds a small LD_PRELOAD utility to catch xattr changes. While for
now it only care about setcap (security.capability), we could also do
that to all extended attributes. But for now, only capabilities are
needed. This small utility saves capability in database indexed by
dev+inode.

For now newuidmap and newgidmap seem to be the only ones that needed
this. But it might be in future less and less setuid will be used and
replaced by setcap.