Add a check to ensure all elements in components are in components stack
Or in `build-depends, depends, runtime-depends` of another element
inside `elements/components/`.

If neither of these two are true then the element is most likely
not getting built in CI and is probably not being updated.

This does not traverse the full deptree (since we are only interested in
elements inside a particular directory) but rather the elements that are
directly present as dependencies in an element inside
`elements/components/` or in a yaml file included in an element inside
`elements/components/`.