sdk: Drop ruby
Ruby is barely used by anything in fdsdk; only two elements uses it
and webkitgtk uses it as build-depends. It's also not used much on
Flathub.

We don't update it often due to being tarballs yet it occupies a
significant amount of space in the SDK around ~30 MB which is not
justified given the lack of use.