Add a test to make sure the perl version doesn't regress
This installs a generic perl module to app, then the test makes sure
we find the module under the short ABI version path (not the path with
patch version!).

The second test checks if perl got compiled with any path that still
has the full version (with patch version) in it. This would mean that
we need to pass some other configure args to set that path to the
ABI version path.