glslang: Don't build with spirv-tools
spirv-tools is built statically and not shipped and we don't want
to build shared or ship static libs. But in this case glslang itself
cannot be found as dependency because the `glslang-config.cmake`
includes:

```
include(CMakeFindDependencyMacro)
if(ON)
   find_dependency(SPIRV-Tools-opt)
endif()
```

https://github.com/KhronosGroup/glslang/blob/b888ebcee8ffaa59536b5f33277936536249d9be/CMakeLists.txt#L356-L365

But with no spirv-tools in SDK, glslang itself cannot be found by other
projects using `glslang`. So disable spirv-tools support and make this
a no-op.

See https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/issues/1874