gnupg: Skip DBUS_SESSION_BUS_ADDRESS inside Flatpak
If gpg is running inside Flatpak there is no guarantee that the X11,
Wayland or D-Bus socket inside the container corresponds 1:1 to the
analogous socket that is available outside.

This used to work by coincidence inside Flatpak, but the addition of
the "Wayland security context" mechanism in 1.15.x resulted in
the `WAYLAND_DISPLAY` usually differing from its value on the host. [1]

The above issue was worked around in Flatpak [2] but an even earlier
issue was found where `DBUS_SESSION_BUS_ADDRESS` inside the Flatpak
`/run/flatpak/bus` is different from host [3].

Passing these incorrect values in stdenvnames from gpg inside the
Flatpak runtime to the agent running on host, causes the pinentry
dialogue to not show up at all. See [4] for issue reports from
various Flatpak applications about this.

So we skip the `DBUS_SESSION_BUS_ADDRESS` in
`session_env_list_stdenvnames()`. According to smcv, a Flatpak maintainer
likely more values should be blocklisted [5] but we never got the full
list and this change makes the pinentry dialogue work from inside
Flatpak [5].

[1]: https://github.com/flatpak/flatpak/issues/5863
[2]: https://github.com/flatpak/flatpak/pull/6018
[3]: https://github.com/flatpak/flatpak/commit/40510e8ae8fd433df04959938b8e81e582849481#diff-12ca338bd8d02696d700648c2b68705b2f354a228a4b6cbab5cd9e4335f63dabR884
[4]: https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/issues/1767
[5]: https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/merge_requests/26722#note_2556926866