tests: Add a test to ensure mktime is working
This regressed when zoneinfo databases were removed from the Flatpak
SDK and Platform via split-rules. The assumption was that since
Flatpak mounts them from host and sets up all the symlinks, this would
not be needed but apparently Flatpak doesn't do so in the build sandbox
and some software such as `R` requires them for compilation.

So we run the test both inside the build sandbox and again in the
run time sandbox