ibus: Enable vala to build tools
Building from git requires generating the C files in `tools/`
with valac that come pre-generated in the tarball. So vala is needed
but passing `--disable-vala` at the same time causes an erroneous
configuration where the mandatory build part - `tools/` cannot be built
because we are passing `--disable-vala`.

Since it is a build time dependency only there is no reason to not
enable it. This also gives us vala binding files.

See https://github.com/ibus/ibus/issues/2770