Move exclude-systemd split rule to project split rule
Marking files with this split rule by default has no effect and the
vm-only domain is only supposed to be (and is currently) excluded
from Flatpak images (in both fdsdk and g-b-m) which does not require
these files (and none are present currently either).

Making it optional per element split has the side effect that we have
to scan and add it to individual elements whenever something "leaks"
in to Flatpak images.

Fixes https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/issues/1685