components/lvm2-base.bst: Fix path to run dir
Unfortunately, the configure script tries to be a bit too smart and
decides to check if /run exits. If not it configures it to be /var/run
instead. This causes issues on GNOME OS when LVM is used on disks.

Co-authored-by: bbhtt <bbhtt.zn0i8@slmail.me>

Ref. 380aded12c9a8f1f4e26eb978473eaecc236e22c