fdsdk-lic-dedup: Don't cobber filenames
There's no need to clobber license filenames when deduplicating: we can
easily keep the original filename and still symlink to the right places.

The filenames might have meaning. For instance: linux-firmware has a
_ton_ of LICENSE.<module name> files, and even if these licenses contain
the same text (for instance, maybe some manufacturer uses the same
license to cover multiple different categories of firmware for their
devices) it would be destructive to the meaning of these license files
to rename them from LICENSE.<module name> to LICENSE_{1,2,3,...}