fdsdk-lic-dedup: Check for common with realpath
We checked each file to see if it has the common dir as one of its
parent directories. Presumably this was done per-file to exclude files
that are already symlinks to the common dir (otherwise we'd just check
once once we enter the dir, rather than for each file in the dir).
However, we don't try to resolve symlinks first, so that doesn't happen