vm/boot: Work-around dracut-install not finding libraries
Dracut-install finally switched from using the dynamic linker to reading
DT_NEEDED directly. Unfortunately it is missing the libdirs. And it has
them hard-coded. It would be more convenient to use `--libdirs` to dracut.
But dracut does not pass that to dracut-install. So as a work-around
we need to write in ld.so.conf which dracut-install does read.