ovmf: Add static copies of stack cookie files to make ovmf reproducible
https://github.com/tianocore/edk2/commit/cac0955658cb591d4629bf90aaa542a66e25be55
made the stack cookie generation dynamic and on a clean build it
generates new copies of these files which are not reproducible.

Apparently we are supposed to reuse them
https://github.com/tianocore/edk2/issues/11433#issuecomment-3218759579

The files are copied from the build directory by adding a `false`
to the build step.

Fixes https://gitlab.com/freedesktop-sdk/freedesktop-sdk/-/issues/1893